Move over release channel
